pdf_QA_agent/requirements.txt
langchain
openai
langchainhub
langchain-openai
chromadb
tiktoken
pypdf
python-dotenv
streamlit